Sanitized all user input - login.php, register.php, user-account.php.
Added a cartTotalQuantity session to display the number of items in the user's cart in the cart button on the navbarUser. When an item is deleted from the cart, the number is reduced by the number of items removed. This resets to 0 when a transaction is processed.

Formatted an empty table for when the cart is empty.

Corrected an error on user-account.php as it was overwriting the password of the user as i used 'placeholder' instead of 'value'.

Changed all page headers to align text left, instead of center, for cohesiveness.